pyproject.toml
[project]
name = "bookapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
]
